+++ lib/processor/text_test.go
@@ -585,6 +585,20 @@ func TestTextReplaceRegexp(t *testing.T) {
 			input:  `baz bar`,
 			output: `baz bar`,
 		},
+		{
+			name:   "replace regexp submatch 1",
+			arg:    "(foo?) (bar?) (baz?)",
+			value:  "hello $2 world",
+			input:  `foo bar baz`,
+			output: `hello bar world`,
+		},
+		{
+			name:   "replace regexp submatch 2",
+			arg:    "(foo?) (bar?) (baz?)",
+			value:  "hello $4 world",
+			input:  `foo bar baz`,
+			output: `hello  world`,
+		},
 	}
 
 	for _, test := range tests {
@@ -619,7 +633,6 @@ func TestTextFindRegexp(t *testing.T) {
 	type jTest struct {
 		name   string
 		arg    string
-		value  string
 		input  string
 		output string
 	}
@@ -649,41 +662,12 @@ func TestTextFindRegexp(t *testing.T) {
 			input:  `baz bar`,
 			output: ``,
 		},
-		{
-			name:   "find regexp submatch 1",
-			arg:    "(foo?)",
-			value:  "1",
-			input:  `foo bar`,
-			output: `foo`,
-		},
-		{
-			name:   "find regexp submatch 2",
-			arg:    "(foo?) (bar?)",
-			value:  "2",
-			input:  `foo bar`,
-			output: `bar`,
-		},
-		{
-			name:   "find regexp submatch 3",
-			arg:    "(foo?) (bar?)",
-			value:  "0",
-			input:  `foo bar`,
-			output: `foo bar`,
-		},
-		{
-			name:   "find regexp submatch 4",
-			arg:    "(foo?) (bar?)",
-			value:  ",,,2,2,,",
-			input:  `foo bar`,
-			output: `barbar`,
-		},
 	}
 
 	for _, test := range tests {
 		conf := NewConfig()
 		conf.Text.Operator = "find_regexp"
 		conf.Text.Arg = test.arg
-		conf.Text.Value = test.value
 		conf.Text.Parts = []int{0}
 
 		tp, err := NewText(conf, nil, log.Noop(), metrics.Noop())